omp_server/dev_requirement.txt
altgraph==0.17.4
amqp==5.0.6
asgiref==3.4.1
aspy.yaml==1.3.0
attrs==21.2.0
backports.entry-points-selectable==1.1.0
bcrypt==3.2.0
billiard==3.6.4.0
boto3==1.16.42
botocore==1.19.42
cached-property==1.5.2
celery==5.1.2
certifi==2021.5.30
cffi==1.14.4
cfgv==3.3.1
chardet==3.0.4
charset-normalizer==2.0.5
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
concurrent-log-handler==0.9.19
contextvars==2.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.1
distlib==0.3.2
distro==1.5.0
Django==3.1.4
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-dmPython==3.1.7
django-filter==2.4.0
django-ipware==4.0.0
django-timezone-field==4.2.1
djangorestframework==3.12.4
djangorestframework-bulk==0.2.1
djangorestframework-jwt==1.11.0
dmPython==2.4.5
emoji==1.5.0
et-xmlfile==1.1.0
filelock==3.0.12
greenlet==2.0.2
identify==2.2.14
idna==3.2
immutables==0.20
importlib-metadata==7.0.0
iniconfig==1.1.1
ipaddress==1.0.23
itypes==1.2.0
jdcal==1.4.1
Jinja2==3.0.1
jmespath==0.10.0
kombu==5.1.0
looseversion==1.3.0
MarkupSafe==2.0.1
msgpack==1.0.2
nodeenv==1.6.0
numpy==1.24.4
openpyxl==3.1.2
packaging==23.1
pandas==2.0.3
paramiko==2.7.2
Pillow==9.1.0
pip==20.3.3
platformdirs==2.3.0
pluggy==1.0.0
portalocker==2.3.0
pre-commit==2.15.0
prettytable==3.6.0
prompt-toolkit==3.0.20
psutil==5.8.0
py==1.10.0
pycparser==2.20
pycryptodome==3.10.1
pycryptodomex==3.9.9
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2023.10
PyJWT==1.7.1
PyMySQL==1.0.2
PyNaCl==1.4.0
pyparsing==2.4.7
pysqlite3==0.4.5
pytest==6.2.5
python-crontab==2.5.1
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
pyzmq==20.0.0
redis==3.5.3
requests==2.26.0
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
s3transfer==0.3.3
salt==3006.3
scp==0.13.3
setuptools==49.2.1
six==1.16.0
SQLAlchemy==2.0.18
sqlalchemy-dm==1.1.10
sqlparse==0.4.1
toml==0.10.2
typing-extensions==4.7.1
tzdata==2023.3
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.26.6
uWSGI==2.0.19.1
vine==5.0.0
virtualenv==20.7.2
wcwidth==0.2.5
Werkzeug==1.0.1
xlrd==1.2.0
zipp==3.17.0